refactor async response payload test
